+++ lib/src/cose.dart
@@ -1,199 +1,139 @@
 // Copyright (c) 2021, Jeroen trappers. All rights reserved. Use of this source
 // code is governed by the license that can be found in the LICENSE file
 
-library dart_cose;
-
-import 'dart:typed_data';
 import 'dart:convert';
+import 'dart:typed_data';
+import 'package:asn1lib/asn1lib.dart';
 import 'package:cbor/cbor.dart';
 import 'package:crypto/crypto.dart';
 import 'package:crypto_keys/crypto_keys.dart';
+import 'package:dart_cose/dart_cose.dart';
+import 'package:dart_cose/src/logger/cose_logger.dart';
+import 'package:dart_cose/src/util/certificate_util.dart';
+import 'package:dart_cose/src/util/header_util.dart';
 import 'package:x509b/x509.dart'; // TODO replace by x509 when upstream fixes are merged
 import 'package:ninja/ninja.dart' as ninja;
 import 'package:ninja/padder/mgf/mgf.dart';
 import 'package:ninja/asymmetric/rsa/encoder/emsaPkcs1v15.dart';
 
-const begin_cert = '-----BEGIN CERTIFICATE-----';
-const end_cert = '-----END CERTIFICATE-----';
-
-const HeaderParameters = {
-  'partyUNonce': -22,
-  'static_key_id': -3,
-  'static_key': -2,
-  'ephemeral_key': -1,
-  'alg': 1,
-  'crit': 2,
-  'content_type': 3,
-  'kid': 4,
-  'IV': 5,
-  'Partial_IV': 6,
-  'counter_signature': 7
-};
-
-enum CoseErrorCode {
-  none,
-  cbor_decoding_error,
-  unsupported_format,
-  invalid_format,
-  unsupported_header_format,
-  invalid_header_format,
-  payload_format_error,
-  key_not_found,
-  kid_mismatch,
-  unsupported_algorithm
-}
-
-class CoseResult {
-  final Map payload;
-  final bool verified;
-  final CoseErrorCode errorCode;
-
-  CoseResult(
-      {required this.payload, required this.verified, required this.errorCode});
-}
-
 class Cose {
+  static const _CBOR_DATA_LENGTH = 4;
+  static const _CBOR_DATA_PROTECTED_HEADER_INDEX = 0;
+  static const _CBOR_DATA_UNPROTECTED_HEADER_INDEX = 1;
+  static const _CBOR_DATA_PAYLOAD_BYTES_INDEX = 2;
+  static const _CBOR_DATA_SIGNER_INDEX = 3;
+
   // input: cose as binary, certs as map kid -> PEM
   static CoseResult decodeAndVerify(List<int> cose, Map<String, String> certs) {
     var inst = Cbor();
     inst.decodeFromList(cose);
     List<dynamic>? data = inst.getDecodedData();
 
-    if (null == data) {
-      return CoseResult(
-          payload: {},
-          verified: false,
-          errorCode: CoseErrorCode.cbor_decoding_error);
-    }
-
-    if (data.isEmpty) {
-      return CoseResult(
-          payload: {},
-          verified: false,
-          errorCode: CoseErrorCode.cbor_decoding_error);
+    //check if the data is not there
+    if ((null == data) || (data.isEmpty)) {
+      return CoseResult.withErrorCode(CoseErrorCode.cbor_decoding_error);
     }
 
     // take the first element
-    var element = data.first;
+    final element = data.first;
 
     // check if it is of type List
     if (!(element is List)) {
-      return CoseResult(
-          payload: {},
-          verified: false,
-          errorCode: CoseErrorCode.unsupported_format);
+      return CoseResult.withErrorCode(CoseErrorCode.unsupported_format);
     }
 
     List items = element;
     // check if it has exactly 4 items
-    if (4 != items.length) {
-      return CoseResult(
-          payload: {},
-          verified: false,
-          errorCode: CoseErrorCode.invalid_format);
+    if (items.length != _CBOR_DATA_LENGTH) {
+      return CoseResult.withErrorCode(CoseErrorCode.invalid_format);
     }
 
     // extract the useful information.
-    final protectedHeader = items[0];
-    final unprotectedHeader = items[1];
-    final payloadBytes = items[2];
-    final signers = items[3];
+    final protectedHeader = items[_CBOR_DATA_PROTECTED_HEADER_INDEX];
+    final unprotectedHeader = items[_CBOR_DATA_UNPROTECTED_HEADER_INDEX];
+    final payloadBytes = items[_CBOR_DATA_PAYLOAD_BYTES_INDEX];
+    final signers = items[_CBOR_DATA_SIGNER_INDEX];
 
     // parse headers.
-    var headers = Cbor();
+    final headers = Cbor();
     headers.decodeFromBuffer(protectedHeader);
     var headerList = headers.getDecodedData();
-    var header = {};
+    var header = <dynamic, dynamic>{};
     if (headerList != null) {
       if (!(headerList is List)) {
-        return CoseResult(
-            payload: {},
-            verified: false,
-            errorCode: CoseErrorCode.unsupported_header_format);
+        return CoseResult.withErrorCode(
+            CoseErrorCode.unsupported_header_format);
       }
 
       if (headerList.isEmpty) {
-        return CoseResult(
-            payload: {},
-            verified: false,
-            errorCode: CoseErrorCode.cbor_decoding_error);
+        return CoseResult.withErrorCode(CoseErrorCode.cbor_decoding_error);
       }
       header = headerList.first;
     }
 
-    final kidKey = HeaderParameters['kid'];
-    // fall back to unprotected header if protected is not provided.
-    var kidBuffer = header[kidKey] ?? unprotectedHeader[kidKey];
-    var kid = Uint8List.view(kidBuffer.buffer, 0, kidBuffer.length);
-    if (kid.length > 8) {
-      kid = kid.sublist(0, 8);
-    }
-    final bkid = base64.encode(kid);
+    final bKid = HeaderUtil.parseKid(header, unprotectedHeader);
+    final a = HeaderUtil.parseAlg(header, unprotectedHeader);
 
-    final algKey = HeaderParameters['alg'];
-    final a = header[algKey] ?? unprotectedHeader[algKey];
-
-    //print("kid: ${base64.encode(kid)}");
-    //print("alg: $a");
+    CoseLogger.print("kid: $bKid");
+    CoseLogger.print("alg: $a");
 
     // parse the payload
     var payloadCbor = Cbor();
     payloadCbor.decodeFromBuffer(payloadBytes);
-    //print(payloadCbor.decodedPrettyPrint());
+    CoseLogger.print(payloadCbor.decodedPrettyPrint());
 
     dynamic payload = {};
     try {
       var data = payloadCbor.getDecodedData();
       if (null == data) {
-        return CoseResult(
-            payload: {},
-            verified: false,
-            errorCode: CoseErrorCode.payload_format_error);
+        return CoseResult.withErrorCodeAndKid(
+            CoseErrorCode.payload_format_error, bKid);
       }
       payload = data.first;
     } on Exception catch (e) {
-      print(e);
-      return CoseResult(
-          payload: {},
-          verified: false,
-          errorCode: CoseErrorCode.payload_format_error);
+      CoseLogger.printError(e);
+      return CoseResult.withErrorCodeAndKid(
+          CoseErrorCode.payload_format_error, bKid);
     }
-    if (!certs.containsKey(bkid)) {
+    if (!certs.containsKey(bKid)) {
       return CoseResult(
           payload: payload,
           verified: false,
-          errorCode: CoseErrorCode.key_not_found);
+          errorCode: CoseErrorCode.key_not_found,
+          coseKid: bKid,
+          certificate: null,
+          publicKey: null);
     }
 
-    String derB64 = certs[bkid]!;
-    String cert = derB64.trim();
+    // Get the public key to verify the signature.
+    // This can be either a x509 certificate (EU) or only the public key structure (UK)
+    // First we try to parse a x509, when that fails we try to treat is as a public key structure
+    PublicKey publicKey;
+    X509Certificate? x509cert;
+    try {
+      final pem = certs[bKid]!;
+      x509cert = CertificateUtil.getX509Certificate(pem);
+      final certKid = extractKid(pem);
+      if (certKid != bKid) {
+        return CoseResult(
+            payload: payload,
+            verified: false,
+            errorCode: CoseErrorCode.kid_mismatch,
+            coseKid: bKid,
+            certificate: x509cert,
+            publicKey: null);
+      }
 
-    // add pem header and footer if missing.
-
-    if (!(cert.startsWith(begin_cert) && cert.endsWith(end_cert))) {
-      cert = begin_cert + '\n' + cert + '\n' + end_cert;
-    } else {
-      derB64 = cert.replaceFirst(begin_cert, "");
-      derB64 = derB64.replaceFirst(end_cert, "");
-      derB64 = derB64.replaceAll("\n", "");
-      derB64 = derB64.replaceAll(" ", "");
+      publicKey = x509cert.publicKey;
+    } on Error {
+      final key = certs[bKid]!;
+      SubjectPublicKeyInfo publicKeyInfo = SubjectPublicKeyInfo.fromAsn1(
+          ASN1Sequence.fromBytes(base64Decode(key)));
+      x509cert = null;
+      publicKey = publicKeyInfo.subjectPublicKey;
     }
 
-    // we expect there to be only 1 cert in the pem, so we take the first.
-    var x509cert = parsePem(cert).first as X509Certificate;
-
-    //The kid is defined as the first 8 bytes of the SHA256 hash of the certificate.
-    var der = base64Decode(derB64);
-    var certKid = base64Encode(sha256.convert(der).bytes.sublist(0, 8));
-
-    if (certKid != bkid) {
-      return CoseResult(
-          payload: payload,
-          verified: false,
-          errorCode: CoseErrorCode.kid_mismatch);
-    }
-
-    var sigStructure = Cbor();
+    final sigStructure = Cbor();
     final sigStructureEncoder = sigStructure.encoder;
 
     sigStructureEncoder.writeArray([
@@ -206,8 +146,6 @@ class Cose {
 
     sigStructure.decodeFromInput();
     final sigStructureBytes = sigStructure.output.getData();
-
-    var publicKey = x509cert.publicKey;
 
     // -7: {'sign': 'ES256', 'digest': 'SHA-256'},
     Verifier? verifier;
@@ -225,42 +163,33 @@ class Cose {
         return CoseResult(
             payload: payload,
             verified: false,
-            errorCode: CoseErrorCode.unsupported_algorithm);
+            errorCode: CoseErrorCode.unsupported_algorithm,
+            coseKid: bKid,
+            certificate: x509cert,
+            publicKey: publicKey);
       }
     } else if (publicKey is RsaPublicKey) {
       // secondary algorithm
       /// RSASSA-PKCS1-v1_5 using SHA-256
       if (-7 == a) {
         verifier = publicKey.createVerifier(algorithms.signing.rsa.sha256);
-        ninja.RsaVerifier ninv =
-            ninja.RsassaPkcs1v15Verifier(hasher: EmsaHasher.sha256);
-        var npk = ninja.RSAPublicKey(publicKey.modulus, publicKey.exponent);
-        var verified = ninv.verify(
-            npk,
-            Uint8List.view(signers.buffer, 0, signers.length),
-            sigStructureBytes.buffer.asUint8List());
-        print(verified);
+        //ninja.RsaVerifier ninv = ninja.RsassaPkcs1v15Verifier(hasher: EmsaHasher.sha256);
+        //var npk = ninja.RSAPublicKey(publicKey.modulus, publicKey.exponent);
+        //var verified = ninv.verify(npk, Uint8List.view(signers.buffer, 0, signers.length), sigStructureBytes.buffer.asUint8List());
+        //CoseLogger.print(verified);
       } else if (-35 == a) {
         verifier = publicKey.createVerifier(algorithms.signing.rsa.sha384);
-        ninja.RsaVerifier ninv =
-            ninja.RsassaPkcs1v15Verifier(hasher: EmsaHasher.sha384);
-        var npk = ninja.RSAPublicKey(publicKey.modulus, publicKey.exponent);
-        var verified = ninv.verify(
-            npk,
-            Uint8List.view(signers.buffer, 0, signers.length),
-            sigStructureBytes.buffer.asUint8List());
-        print(verified);
+        //ninja.RsaVerifier ninv = ninja.RsassaPkcs1v15Verifier(hasher: EmsaHasher.sha384);
+        //var npk = ninja.RSAPublicKey(publicKey.modulus, publicKey.exponent);
+        //var verified = ninv.verify(npk, Uint8List.view(signers.buffer, 0, signers.length), sigStructureBytes.buffer.asUint8List());
+        //CoseLogger.print(verified);
       } else if (-36 == a) {
         verifier = publicKey.createVerifier(algorithms.signing.rsa.sha512);
 
-        ninja.RsaVerifier ninv =
-            ninja.RsassaPkcs1v15Verifier(hasher: EmsaHasher.sha512);
-        var npk = ninja.RSAPublicKey(publicKey.modulus, publicKey.exponent);
-        var verified = ninv.verify(
-            npk,
-            Uint8List.view(signers.buffer, 0, signers.length),
-            sigStructureBytes.buffer.asUint8List());
-        print(verified);
+        //ninja.RsaVerifier ninv = ninja.RsassaPkcs1v15Verifier(hasher: EmsaHasher.sha512);
+        //var npk = ninja.RSAPublicKey(publicKey.modulus, publicKey.exponent);
+        //var verified = ninv.verify(npk, Uint8List.view(signers.buffer, 0, signers.length), sigStructureBytes.buffer.asUint8List());
+        //CoseLogger.print(verified);
       } else if (-37 == a) {
         ninja.RsaSsaPssVerifier ninv = ninja.RsaSsaPssVerifier(
             hasher: sha256, mgf: Mgf1(hasher: sha256), saltLength: 32);
@@ -270,19 +199,24 @@ class Cose {
             npk,
             Uint8List.view(signers.buffer, 0, signers.length),
             sigStructureBytes.buffer.asUint8List());
-
-        print(verified);
+        CoseLogger.print(verified);
       } else {
         return CoseResult(
             payload: payload,
             verified: false,
-            errorCode: CoseErrorCode.unsupported_algorithm);
+            errorCode: CoseErrorCode.unsupported_algorithm,
+            coseKid: bKid,
+            certificate: x509cert,
+            publicKey: publicKey);
       }
     } else {
       return CoseResult(
           payload: payload,
           verified: false,
-          errorCode: CoseErrorCode.unsupported_algorithm);
+          errorCode: CoseErrorCode.unsupported_algorithm,
+          coseKid: bKid,
+          certificate: x509cert,
+          publicKey: publicKey);
     }
 
     if (!verified && verifier != null) {
@@ -291,6 +225,13 @@ class Cose {
     }
 
     return CoseResult(
-        payload: payload, verified: verified, errorCode: CoseErrorCode.none);
+        payload: payload,
+        verified: verified,
+        errorCode: CoseErrorCode.none,
+        coseKid: bKid,
+        certificate: x509cert,
+        publicKey: publicKey);
   }
+
+  static String extractKid(String pem) => CertificateUtil.extractKid(pem);
 }